+++ RemeApp/Controller/SalesFloorMapView /SalesFloorMapViewController.swift
@@ -187,7 +187,7 @@ class SalesFloorMapViewController: UIViewController {
                        greenFourButton, greenFiveButton]
         // for文でbuttonsに順番にアクセス
         for (index, button) in buttons.enumerated() {
-            let salesFloor = SalesFloorType(rawValue: index)!
+            let salesFloor = DefaultSalesFloorType(rawValue: index)!
             button?.setTitle(salesFloor.nameOfSalesFloor, for: .normal)
             if (errandDataList.first(where: {$0.salesFloorRawValue == index})?.salesFloorRawValue) != nil {
                 button?.backgroundColor = salesFloor.colorOfSalesFloor
