+++ RemeApp/Extension/ExtensionUIImage.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 
+// UIImageをリサイズできるように拡張
 extension UIImage {
     func resize(to size: CGSize) -> UIImage? {
         UIGraphicsBeginImageContextWithOptions(size, false, 0.0)
